TPC SCD outlier filtering refactoring
Outlier filtering is moved from TrackResiduals to TrackInterpolation
class in order to allow for setting the maximum number of successfully
processed tracks.
Some functionality is removed which allowed to process converted residuals
from Run 2. This adds some cleaner separation. TPC cluster residuals are
collected in TrackInterpolation. TrackResiduals is for creating the
average map from these residuals.
Also added some reader/writer for unbinned residuals to run the aggregator
in a standalone mode.
