+++ Detectors/GlobalTrackingWorkflow/tpcinterpolationworkflow/src/tpc-interpolation-workflow.cxx
@@ -39,7 +39,7 @@ void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
     {"disable-mc", VariantType::Bool, false, {"disable MC propagation even if available"}},
     {"enable-itsonly", VariantType::Bool, false, {"process tracks without outer point (ITS-TPC only)"}},
     {"tracking-sources", VariantType::String, std::string{GID::ALL}, {"comma-separated list of sources to use for tracking"}},
-    {"enable-residual-writer", VariantType::Bool, false, {"write unbinned and unfiltered residuals"}},
+    {"write-unfiltered", VariantType::Bool, false, {"write unfiltered residuals"}},
     {"send-track-data", VariantType::Bool, false, {"Send also the track information to the aggregator"}},
     {"configKeyValues", VariantType::String, "", {"Semicolon separated key=value strings ..."}}};
   o2::raw::HBFUtilsInitializer::addConfigOption(options);
@@ -74,14 +74,14 @@ WorkflowSpec defineDataProcessing(ConfigContext const& configcontext)
   auto useMC = !configcontext.options().get<bool>("disable-mc");
   useMC = false; // force disabling MC as long as it is not implemented
   auto processITSTPConly = configcontext.options().get<bool>("enable-itsonly");
-  auto writeResiduals = configcontext.options().get<bool>("enable-residual-writer");
+  auto writeUnfilteredResiduals = configcontext.options().get<bool>("write-unfiltered");
   auto sendTrackData = configcontext.options().get<bool>("send-track-data");
   GID::mask_t src = allowedSources & GID::getSourcesMask(configcontext.options().get<std::string>("tracking-sources"));
   LOG(info) << "Data sources: " << GID::getSourcesNames(src);
 
-  specs.emplace_back(o2::tpc::getTPCInterpolationSpec(src, useMC, processITSTPConly, writeResiduals, sendTrackData));
-  if (!configcontext.options().get<bool>("disable-root-output") && writeResiduals) {
-    specs.emplace_back(o2::tpc::getTPCResidualWriterSpec(useMC));
+  specs.emplace_back(o2::tpc::getTPCInterpolationSpec(src, useMC, processITSTPConly, writeUnfilteredResiduals, sendTrackData));
+  if (!configcontext.options().get<bool>("disable-root-output")) {
+    specs.emplace_back(o2::tpc::getTPCResidualWriterSpec(writeUnfilteredResiduals, sendTrackData));
   }
 
   o2::globaltracking::InputHelper::addInputSpecs(configcontext, specs, src, src, src, useMC);
