+++ Detectors/TPC/calibration/SpacePoints/include/SpacePoints/TrackResiduals.h
@@ -104,46 +104,12 @@ class TrackResiduals
     ClassDefNV(LocalResid, 1);
   };
 
-  /// Structure for TPC residuals before the binning, but after outlier rejection
-  /// This is the data type which will be send from the EPNs to the aggregator
-  struct UnbinnedResid {
-    UnbinnedResid() = default;
-    UnbinnedResid(float dyIn, float dzIn, float tgSlpIn, float yIn, float zIn, unsigned char rowIn, unsigned char secIn) : dy(static_cast<short>(dyIn * 0x7fff / param::MaxResid)),
-                                                                                                                           dz(static_cast<short>(dzIn * 0x7fff / param::MaxResid)),
-                                                                                                                           tgSlp(static_cast<short>(tgSlpIn * 0x7fff / param::MaxTgSlp)),
-                                                                                                                           y(static_cast<short>(yIn * 0x7fff / param::MaxY)),
-                                                                                                                           z(static_cast<short>(zIn * 0x7fff / param::MaxZ)),
-                                                                                                                           row(rowIn),
-                                                                                                                           sec(secIn) {}
-    short dy;          ///< residual in y
-    short dz;          ///< residual in z
-    short tgSlp;       ///< track dip angle
-    short y;           ///< y position of the track, needed for binning
-    short z;           ///< z position of the track, needed for binning
-    unsigned char row; ///< TPC pad row
-    unsigned char sec; ///< TPC sector (0..35)
-    ClassDefNV(UnbinnedResid, 1);
-  };
-
   /// Structure which holds the statistics for each voxel
   struct VoxStats {
     VoxStats() = default;
     std::array<float, VoxDim> meanPos{};
     float nEntries{0.f};
     ClassDefNV(VoxStats, 1);
-  };
-
-  /// Structure for on-the-fly re-calculated track parameters at the validation stage
-  struct TrackParams {
-    TrackParams() = default;
-    float qpt{0.f};
-    float tgl{0.f};
-    std::array<float, param::NPadRows> zTrk{};
-    std::array<float, param::NPadRows> xTrk{};
-    std::array<float, param::NPadRows> dy{};
-    std::array<float, param::NPadRows> dz{};
-    std::array<float, param::NPadRows> tglArr{};
-    std::bitset<param::NPadRows> flagRej{};
   };
 
   // -------------------------------------- initialization --------------------------------------------------
@@ -177,61 +143,12 @@ class TrackResiduals
 
   // -------------------------------------- I/O --------------------------------------------------
 
-  /// Load input from file
-  void loadInputFromFile();
-
   std::vector<LocalResid>& getLocalResVec() { return mLocalResidualsIn; }
   std::vector<VoxStats>** getVoxStatPtr() { return &mVoxStatsInPtr; }
 
   const std::array<std::vector<VoxRes>, SECTORSPERSIDE * SIDES>& getVoxelResults() const { return mVoxelResults; }
 
-  /// Set input directly
-  /// \param trkData Vector with seeding track information (contains range ref to related TPC cluster residuals)
-  /// \param clResiduals Vector of all TPC cluster residuals
-  void setInputData(std::vector<TrackData>& trkData, std::vector<TPCClusterResiduals>& clResiduals);
-
-  /// Returns the collected unbinned residuals after outlier rejection is applied
-  std::vector<UnbinnedResid>& getUnbinnedResiduals() { return mUnbinnedResiduals; }
-
-  /// Returns the track parameters and cluster range reference for validated tracks (this is a subset of the input mTrackDataPtr)
-  std::vector<TrackData>& getTrackDataOut() { return mTrackDataOut; }
-
-  /// Create output files for each sector with trees for local residuals
-  void prepareLocalResidualTrees();
-
-  /// Write trees with local residuals to file
-  void writeLocalResidualTreesToFile();
-
-  /// Clear the vector of unbinned residuals
-  void resetUnbinnedResiduals();
-
-  /// Clear the vector with track data belonging to residuals
-  void resetTrackData() { mTrackDataOut.clear(); }
-
   // -------------------------------------- steering functions --------------------------------------------------
-
-  /// Loads residual data from track interpolation and fills vector of unbinned residuals
-  /// This function also steers the outlier rejection
-  /// \param writeToFile Flag if the local residuals and voxel statistics should be dumped to a file
-  void doOutlierRejection(bool writeToFile = false);
-
-  /// Validates the given input track and its residuals
-  /// \param trk The track parameters, e.g. q/pT, eta, ...
-  /// \param params Structure with per pad information recalculated on the fly
-  /// \return true if the track could be validated, false otherwise
-  bool validateTrack(const TrackData& trk, TrackParams& params) const;
-
-  /// Filter out individual outliers from all cluster residuals of given track
-  /// \return true for tracks which pass the cuts on e.g. max. masked clusters and false for rejected tracks
-  bool outlierFiltering(const TrackData& trk, TrackParams& params) const;
-
-  /// Is called from outlierFiltering() and does the actual calculations (moving average filter etc.)
-  /// \return The RMS of the long range moving average
-  float checkResiduals(const TrackData& trk, TrackParams& params) const;
-
-  /// Calculates the differences in Y and Z for a given set of clusters to a fitted helix.
-  /// First a circular fit in the azimuthal plane is performed and subsequently a linear fit in the transversal plane
-  bool compareToHelix(const TrackData& trk, TrackParams& params) const;
 
   /// Processes residuals for given sector.
   /// \param iSec Sector to process
@@ -333,19 +250,13 @@ class TrackResiduals
   /// \param yc fit result for circle center position in y is stored here
   /// \param r fit result for circle radius is stored here
   /// \param residHelixY residuals in y from fitted circle to given points is stored here
-  void fitCircle(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, float& xc, float& yc, float& r, std::array<float, param::NPadRows>& residHelixY) const;
+  static void fitCircle(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, float& xc, float& yc, float& r, std::array<float, param::NPadRows>& residHelixY);
 
   /// Fits a straight line to a given set of points, w/o taking into account measurement errors or different weights for the points
   /// Straight line is given by y = a * x + b
   /// \param res[0] contains the slope (a)
   /// \param res[1] contains the offset (b)
-  bool fitPoly1(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, std::array<float, 2>& res) const;
-
-  /// For a given set of points, calculate the differences from each point to the fitted lines from all other points in their neighbourhoods (+- mNMAShort points)
-  void diffToLocLine(const int np, int idxOffset, const std::array<float, param::NPadRows>& x, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffY) const;
-
-  /// For a given set of points, calculate their deviation from the moving average (build from the neighbourhood +- mNMALong points)
-  void diffToMA(const int np, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffMA) const;
+  static bool fitPoly1(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, std::array<float, 2>& res);
 
   // -------------------------------------- binning / geometry --------------------------------------------------
 
@@ -496,9 +407,6 @@ class TrackResiduals
 
   // -------------------------------------- settings --------------------------------------------------
 
-  void setLocalResFileName(std::string fName) { mLocalResFileName = fName; }
-  void setLocalResTreeName(std::string tName) { mLocalResTreeName = tName; }
-  void setLocalResBranchName(std::string bName) { mLocalResBranchName = bName; }
   void setMaxPointsPerSector(int nPoints) { mMaxPointsPerSector = nPoints; }
   void setMinEntriesPerVoxel(int nEntries) { mMinEntriesPerVoxel = nEntries; }
   void setLTMCut(float ltmCut) { mLTMCut = ltmCut; }
@@ -514,9 +422,6 @@ class TrackResiduals
   void setMaxSigZ(float sigZ) { mMaxSigZ = sigZ; }
   void setMaxGaussStdDev(float sigmas) { mMaxGaussStdDev = sigmas; }
 
-  std::string getLocalResFileName() const { return mLocalResFileName; }
-  std::string getLocalResTreeName() const { return mLocalResTreeName; }
-  std::string getLocalResBranchName() const { return mLocalResBranchName; }
   int getMaxPointsPerSector() const { return mMaxPointsPerSector; }
   int getMinEntriesPerVoxel() const { return mMinEntriesPerVoxel; }
   float getLTMCut() const { return mLTMCut; }
@@ -548,8 +453,6 @@ class TrackResiduals
   void closeOutputFile();
 
  private:
-  // names of input files / trees
-  std::string mInputFileNameResiduals{"residuals_tpc.root"}; ///< name of file with track residuals
   // some constants
   static constexpr float sFloatEps{1.e-7f}; ///< float epsilon for robust linear fitting
   static constexpr float sDeadZone{1.5f};   ///< dead zone for TPC in between sectors
@@ -558,19 +461,11 @@ class TrackResiduals
   static constexpr int sMaxSmtDim{7};       ///< max matrix size for smoothing (pol2)
 
   // input data
-  std::unique_ptr<TFile> mFileIn;                       ///< input file with residuals data
-  TTree* mTreeInTracks{nullptr};                        ///< tree with input track information
-  std::vector<TrackData>* mTrackDataPtr{nullptr};       ///< vector with input track information
-  TTree* mTreeInClRes{nullptr};                         ///< tree with TPC cluster residuals
-  std::vector<TPCClusterResiduals>* mClResPtr{nullptr}; ///< vector with unbinned TPC cluster residuals
   std::vector<LocalResid> mLocalResidualsIn;            ///< binned local residuals from aggregator
   std::vector<VoxStats> mVoxStatsIn, *mVoxStatsInPtr{&mVoxStatsIn}; ///< the statistics information for each voxel from the aggregator
   // output data
   std::unique_ptr<TFile> mFileOut; ///< output debug file
   std::unique_ptr<TTree> mTreeOut; ///< tree holding debug output
-  std::vector<UnbinnedResid> mUnbinnedResiduals; ///< large vector for the unbinned residual data which is sent to the aggregator
-  std::vector<UnbinnedResid>* mUnbinnedResidualsPtr{&mUnbinnedResiduals};
-  std::vector<TrackData> mTrackDataOut; // the same as mTrackDataPtr, but the rejected tracks are removed, to be sent to the aggregator
   // status flags
   bool mIsInitialized{}; ///< initialize only once
   bool mPrintMem{};      ///< turn on to print memory usage at certain points
@@ -594,14 +489,8 @@ class TrackResiduals
   std::vector<float> mZ2XBinsCenter{};     ///< bin center in z/x within the interval [0..1]
   float mMaxZ2X{1.f};                      ///< max z/x value
   std::array<bool, VoxDim> mUniformBins{true, true, true}; ///< if binning is uniform for each dimension
-  // local residual data, extracted from track interpolation
-  std::unique_ptr<TFile> mTmpFile{}; ///< I/O file
-  std::unique_ptr<TTree> mTmpTree{}; ///< I/O tree
   // settings
   float mBz{param::Bz};                          ///< nominal magnetic field along Z
-  std::string mLocalResFileName{"deltasSect"};   ///< filename for local residuals input
-  std::string mLocalResTreeName{"treeSec"};      ///< name for tree with local residuals
-  std::string mLocalResBranchName{"resids"};     ///< branch with LocalResid objects
   int mMaxPointsPerSector{30'000'000};           ///< maximum number of accepted points per sector
   int mMinEntriesPerVoxel{15};                   ///< minimum number of points in voxel for processing
   float mLTMCut{.75f};                           ///< fraction op points to keep when trimming input data
@@ -631,11 +520,6 @@ class TrackResiduals
   std::array<std::vector<VoxRes>, SECTORSPERSIDE * SIDES> mVoxelResults{};                  ///< results per sector and per voxel for 3-D distortions
   VoxRes mVoxelResultsOut{};                                                                ///< the results from mVoxelResults are copied in here to be able to stream them
   VoxRes* mVoxelResultsOutPtr{&mVoxelResultsOut};                                           ///< pointer to set the branch address to for the output
-  bool mFilterOutliers = true;                             ///< flag, if outliers from the cluster residual trees should be rejected
-  int mNMALong{15}; ///< number of points to be used for moving average (long range)
-  int mNMAShort{3}; ///< number of points to be used for estimation of distance from local line (short range)
-  float mMaxRejFrac{.15f}; ///< if the fraction of rejected clusters of a track is higher, the full track is invalidated
-  float mMaxRMSLong{.8f};  ///< maximum variance of the cluster residuals wrt moving avarage for a track to be considered
 
   ClassDefNV(TrackResiduals, 2);
 };
