+++ Detectors/TPC/calibration/SpacePoints/src/TrackInterpolation.cxx
@@ -16,6 +16,7 @@
 ///
 
 #include "SpacePoints/TrackInterpolation.h"
+#include "SpacePoints/TrackResiduals.h"
 #include "TPCBase/ParameterElectronics.h"
 #include "DataFormatsTPC/TrackTPC.h"
 #include "DataFormatsTPC/Defs.h"
@@ -45,7 +46,8 @@ void TrackInterpolation::init()
 
   mFastTransform = std::move(TPCFastTransformHelperO2::instance()->create(0));
 
-  mRecoParam.setBfield(o2::base::Propagator::Instance()->getNominalBz());
+  mBz = o2::base::Propagator::Instance()->getNominalBz();
+  mRecoParam.setBfield(mBz);
   mGeoTRD = o2::trd::Geometry::instance();
 
   mInitDone = true;
@@ -91,6 +93,7 @@ void TrackInterpolation::process(const o2::globaltracking::RecoContainer& inp, c
 void TrackInterpolation::interpolateTrack(int iSeed)
 {
   TrackData trackData;
+  std::vector<TPCClusterResiduals> clusterResiduals;
   auto propagator = o2::base::Propagator::Instance();
   const auto& gidTable = (*mGIDtables)[iSeed];
   const auto& trkTPC = mRecoCont->getTPCTrack(gidTable[GTrackID::TPC]);
@@ -137,7 +140,6 @@ void TrackInterpolation::interpolateTrack(int iSeed)
     mCache[iRow].szy[ExtOut] = trkWork.getSigmaZY();
     mCache[iRow].sz2[ExtOut] = trkWork.getSigmaZ2();
     mCache[iRow].snp[ExtOut] = trkWork.getSnp();
-    mCache[iRow].tgl[ExtOut] = trkWork.getTgl();
     //printf("Track alpha at row %i: %.2f, Y(%.2f), Z(%.2f)\n", iRow, trkWork.getAlpha(), trkWork.getY(), trkWork.getZ());
   }
 
@@ -227,12 +229,10 @@ void TrackInterpolation::interpolateTrack(int iSeed)
     mCache[iRow].szy[ExtIn] = trkWork.getSigmaZY();
     mCache[iRow].sz2[ExtIn] = trkWork.getSigmaZ2();
     mCache[iRow].snp[ExtIn] = trkWork.getSnp();
-    mCache[iRow].tgl[ExtIn] = trkWork.getTgl();
   }
 
   // calculate weighted mean at each pad row (assume for now y and z are uncorrelated) and store residuals to TPC clusters
   unsigned short deltaRow = 0;
-  unsigned short nMeasurements = 0;
   for (int iRow = 0; iRow < param::NPadRows; ++iRow) {
     if (!mCache[iRow].clAvailable) {
       ++deltaRow;
@@ -243,9 +243,8 @@ void TrackInterpolation::interpolateTrack(int iSeed)
     mCache[iRow].y[Int] = (mCache[iRow].y[ExtOut] / mCache[iRow].sy2[ExtOut] + mCache[iRow].y[ExtIn] / mCache[iRow].sy2[ExtIn]) / wTotY;
     mCache[iRow].z[Int] = (mCache[iRow].z[ExtOut] / mCache[iRow].sz2[ExtOut] + mCache[iRow].z[ExtIn] / mCache[iRow].sz2[ExtIn]) / wTotZ;
 
-    // simple average w/o weighting for angles
+    // simple average w/o weighting for angle
     mCache[iRow].snp[Int] = (mCache[iRow].snp[ExtOut] + mCache[iRow].snp[ExtIn]) / 2.f;
-    mCache[iRow].tgl[Int] = (mCache[iRow].tgl[ExtOut] + mCache[iRow].tgl[ExtIn]) / 2.f;
 
     TPCClusterResiduals res;
     res.setDY(mCache[iRow].clY - mCache[iRow].y[Int]);
@@ -253,11 +252,9 @@ void TrackInterpolation::interpolateTrack(int iSeed)
     res.setY(mCache[iRow].y[Int]);
     res.setZ(mCache[iRow].z[Int]);
     res.setSnp(mCache[iRow].snp[Int]);
-    res.setTgl(mCache[iRow].tgl[Int]);
     res.sec = mCache[iRow].clSec;
     res.dRow = deltaRow;
-    mClRes.push_back(std::move(res));
-    ++nMeasurements;
+    clusterResiduals.push_back(std::move(res));
     deltaRow = 1;
   }
 
@@ -274,10 +271,41 @@ void TrackInterpolation::interpolateTrack(int iSeed)
   trackData.nClsITS = trkITS.getNumberOfClusters();
   trackData.nTrkltsTRD = gidTable[GTrackID::TRD].isIndexSet() ? mRecoCont->getITSTPCTRDTrack<o2::trd::TrackTRD>(gidTable[GTrackID::ITSTPCTRD]).getNtracklets() : 0;
   trackData.clAvailTOF = gidTable[GTrackID::TOF].isIndexSet() ? 1 : 0;
-  trackData.clIdx.setEntries(nMeasurements);
 
-  mTrackData.push_back(std::move(trackData));
-  mGIDsSuccess.push_back((*mGIDs)[iSeed]);
+  /*
+  // FIXME
+
+  Calculate number of tracks required per TF based on calibration slot length
+  In case too many tracks available, use std::sample algorithm to take random sample of input tracks
+  (make sure to use first most global tracks, then ITS-TPC-TRD, then ITS-TPC-TOF)
+  */
+
+  TrackParams params; // for refitted track parameters and flagging rejected clusters
+  if (validateTrack(trackData, params, clusterResiduals)) {
+    // track is good
+    int nClValidated = 0;
+    int iRow = 0;
+    for (unsigned int iCl = 0; iCl < clusterResiduals.size(); ++iCl) {
+      iRow += clusterResiduals[iCl].dRow;
+      if (params.flagRej[iCl]) {
+        // skip masked cluster residual
+        continue;
+      }
+      ++nClValidated;
+      float tgPhi = clusterResiduals[iCl].snp / std::sqrt((1.f - clusterResiduals[iCl].snp) * (1.f + clusterResiduals[iCl].snp));
+      mClRes.emplace_back(clusterResiduals[iCl].dy, clusterResiduals[iCl].dz, tgPhi, clusterResiduals[iCl].y, clusterResiduals[iCl].z, iRow, clusterResiduals[iCl].sec);
+    }
+    trackData.clIdx.setEntries(nClValidated);
+    mTrackData.push_back(std::move(trackData));
+    mGIDsSuccess.push_back((*mGIDs)[iSeed]);
+  }
+  if (mWriteUnfiltered) {
+    TrackData trkDataTmp = trackData;
+    trkDataTmp.clIdx.setFirstEntry(mClResUnfiltered.size());
+    trkDataTmp.clIdx.setEntries(clusterResiduals.size());
+    mTrackDataUnfiltered.push_back(std::move(trkDataTmp));
+    mClResUnfiltered.insert(mClResUnfiltered.end(), clusterResiduals.begin(), clusterResiduals.end());
+  }
 }
 
 void TrackInterpolation::extrapolateTrack(int iSeed)
@@ -285,6 +313,7 @@ void TrackInterpolation::extrapolateTrack(int iSeed)
   // extrapolate ITS-only track through TPC and store residuals to TPC clusters in the output vectors
   const auto& gidTable = (*mGIDtables)[iSeed];
   TrackData trackData;
+  std::vector<TPCClusterResiduals> clusterResiduals;
   trackData.clIdx.setFirstEntry(mClRes.size());
   const auto& trkITS = mRecoCont->getITSTrack(gidTable[GTrackID::ITS]);
   const auto& trkTPC = mRecoCont->getTPCTrack(gidTable[GTrackID::TPC]);
@@ -311,11 +340,10 @@ void TrackInterpolation::extrapolateTrack(int iSeed)
     res.setY(trkWork.getY());
     res.setZ(trkWork.getZ());
     res.setSnp(trkWork.getSnp());
-    res.setTgl(trkWork.getTgl());
     res.sec = sector;
     res.dRow = row - rowPrev;
     rowPrev = row;
-    mClRes.push_back(std::move(res));
+    clusterResiduals.push_back(std::move(res));
     ++nMeasurements;
   }
   trackData.gid = (*mGIDs)[iSeed];
@@ -330,14 +358,401 @@ void TrackInterpolation::extrapolateTrack(int iSeed)
   trackData.nClsITS = trkITS.getNumberOfClusters();
   trackData.clIdx.setEntries(nMeasurements);
 
-  mTrackData.push_back(std::move(trackData));
-  mGIDsSuccess.push_back((*mGIDs)[iSeed]);
+  TrackParams params; // for refitted track parameters and flagging rejected clusters
+  if (validateTrack(trackData, params, clusterResiduals)) {
+    // track is good
+    int nClValidated = 0;
+    int iRow = 0;
+    for (unsigned int iCl = 0; iCl < clusterResiduals.size(); ++iCl) {
+      iRow += clusterResiduals[iCl].dRow;
+      if (params.flagRej[iCl]) {
+        // skip masked cluster residual
+        continue;
+      }
+      ++nClValidated;
+      float tgPhi = clusterResiduals[iCl].snp / std::sqrt((1.f - clusterResiduals[iCl].snp) * (1.f + clusterResiduals[iCl].snp));
+      mClRes.emplace_back(clusterResiduals[iCl].dy, clusterResiduals[iCl].dz, tgPhi, clusterResiduals[iCl].y, clusterResiduals[iCl].z, iRow, clusterResiduals[iCl].sec);
+    }
+    trackData.clIdx.setEntries(nClValidated);
+    mTrackData.push_back(std::move(trackData));
+    mGIDsSuccess.push_back((*mGIDs)[iSeed]);
+  }
+  if (mWriteUnfiltered) {
+    TrackData trkDataTmp = trackData;
+    trkDataTmp.clIdx.setFirstEntry(mClResUnfiltered.size());
+    trkDataTmp.clIdx.setEntries(clusterResiduals.size());
+    mTrackDataUnfiltered.push_back(std::move(trkDataTmp));
+    mClResUnfiltered.insert(mClResUnfiltered.end(), clusterResiduals.begin(), clusterResiduals.end());
+  }
+}
+
+bool TrackInterpolation::validateTrack(const TrackData& trk, TrackParams& params, const std::vector<TPCClusterResiduals>& clsRes) const
+{
+  if (clsRes.size() < param::MinNCl) {
+    // no enough clusters for this track to be considered
+    LOG(debug) << "Skipping track with too few clusters: " << clsRes.size();
+    return false;
+  }
+
+  bool resHelix = compareToHelix(trk, params, clsRes);
+  if (!resHelix) {
+    LOG(debug) << "Skipping track too far from helix approximation";
+    return false;
+  }
+  if (fabsf(params.qpt) > param::MaxQ2Pt) {
+    LOG(debug) << "Skipping track with too high q/pT: " << params.qpt;
+    return false;
+  }
+  if (!outlierFiltering(trk, params, clsRes)) {
+    return false;
+  }
+  return true;
+}
+
+bool TrackInterpolation::compareToHelix(const TrackData& trk, TrackParams& params, const std::vector<TPCClusterResiduals>& clsRes) const
+{
+  std::array<float, param::NPadRows> residHelixY;
+  std::array<float, param::NPadRows> residHelixZ;
+
+  std::array<float, param::NPadRows> xLab;
+  std::array<float, param::NPadRows> yLab;
+  std::array<float, param::NPadRows> sPath;
+
+  float curvature = fabsf(trk.p[o2::track::ParLabels::kQ2Pt] * mBz * o2::constants::physics::LightSpeedCm2S * 1e-14f);
+  int secFirst = clsRes[0].sec;
+  float phiSect = (secFirst + .5f) * o2::constants::math::SectorSpanRad;
+  float snPhi = sin(phiSect);
+  float csPhi = cos(phiSect);
+  sPath[0] = 0.f;
+
+  int iRow = 0;
+  int nCl = clsRes.size();
+  for (unsigned int iP = 0; iP < nCl; ++iP) {
+    iRow += clsRes[iP].dRow;
+    float yTrk = clsRes[iP].y;
+    // LOGF(info, "iRow(%i), yTrk(%f)", iRow, yTrk);
+    xLab[iP] = param::RowX[iRow];
+    if (clsRes[iP].sec != secFirst) {
+      float phiSectCurrent = (clsRes[iP].sec + .5f) * o2::constants::math::SectorSpanRad;
+      float cs = cos(phiSectCurrent - phiSect);
+      float sn = sin(phiSectCurrent - phiSect);
+      xLab[iP] = param::RowX[iRow] * cs - yTrk * sn;
+      yLab[iP] = yTrk * cs + param::RowX[iRow] * sn;
+    } else {
+      xLab[iP] = param::RowX[iRow];
+      yLab[iP] = yTrk;
+    }
+    // this is needed only later, but we retrieve it already now to save another loop
+    params.zTrk[iP] = clsRes[iP].z;
+    params.xTrk[iP] = param::RowX[iRow];
+    params.dy[iP] = clsRes[iP].dy;
+    params.dz[iP] = clsRes[iP].dz;
+    // done retrieving values for later
+    if (iP > 0) {
+      float dx = xLab[iP] - xLab[iP - 1];
+      float dy = yLab[iP] - yLab[iP - 1];
+      float ds2 = dx * dx + dy * dy;
+      float ds = sqrt(ds2); // circular path (linear approximation)
+      // if the curvature of the track or the (approximated) chord length is too large the more exact formula is used:
+      // chord length = 2r * asin(ds/(2r))
+      // using the first two terms of the tailer expansion for asin(x) ~ x + x^3 / 6
+      if (ds * curvature > 0.05) {
+        ds *= (1.f + ds2 * curvature * curvature / 24.f);
+      }
+      sPath[iP] = sPath[iP - 1] + ds;
+    }
+  }
+  float xcSec = 0.f;
+  float ycSec = 0.f;
+  float r = 0.f;
+  TrackResiduals::fitCircle(nCl, xLab, yLab, xcSec, ycSec, r, residHelixY);
+  // LOGF(info, "Done with circle fit. nCl(%i), xcSec(%f), ycSec(%f), r(%f).", nCl, xcSec, ycSec, r);
+  /*
+  for (int i=0; i<nCl; ++i) {
+    LOGF(info, "i(%i), xLab(%f), yLab(%f).", i, xLab[i], yLab[i]);
+  }
+  */
+  // determine curvature
+  float phiI = TMath::ATan2(yLab[0], xLab[0]);
+  float phiF = TMath::ATan2(yLab[nCl - 1], xLab[nCl - 1]);
+  if (phiI < 0) {
+    phiI += o2::constants::math::TwoPI;
+  }
+  if (phiF < 0) {
+    phiF += o2::constants::math::TwoPI;
+  }
+  float dPhi = phiF - phiI;
+  float curvSign = -1.f;
+  if (dPhi > 0) {
+    if (dPhi < o2::constants::math::PI) {
+      curvSign = 1.f;
+    }
+  } else if (dPhi < -o2::constants::math::PI) {
+    curvSign = 1.f;
+  }
+  params.qpt = std::copysign(1.f / (r * mBz * o2::constants::physics::LightSpeedCm2S * 1e-14f), curvSign);
+
+  // calculate circle coordinates in the lab frame
+  float xc = xcSec * csPhi - ycSec * snPhi;
+  float yc = xcSec * snPhi + ycSec * csPhi;
+
+  std::array<float, 2> pol1Z;
+  TrackResiduals::fitPoly1(nCl, sPath, params.zTrk, pol1Z);
+
+  params.tgl = pol1Z[0];
+
+  // max deviations in both directions from helix fit in y and z
+  float hMinY = 1e9f;
+  float hMaxY = -1e9f;
+  float hMinZ = 1e9f;
+  float hMaxZ = -1e9f;
+  // extract residuals in Z and fill track slopes in sector frame
+  int secCurr = secFirst;
+  iRow = 0;
+  for (unsigned int iCl = 0; iCl < nCl; ++iCl) {
+    iRow += clsRes[iCl].dRow;
+    float resZ = params.zTrk[iCl] - (pol1Z[1] + sPath[iCl] * pol1Z[0]);
+    residHelixZ[iCl] = resZ;
+    if (resZ < hMinZ) {
+      hMinZ = resZ;
+    }
+    if (resZ > hMaxZ) {
+      hMaxZ = resZ;
+    }
+    if (residHelixY[iCl] < hMinY) {
+      hMinY = residHelixY[iCl];
+    }
+    if (residHelixY[iCl] > hMaxY) {
+      hMaxY = residHelixY[iCl];
+    }
+    int sec = clsRes[iCl].sec;
+    if (sec != secCurr) {
+      secCurr = sec;
+      phiSect = (.5f + sec) * o2::constants::math::SectorSpanRad;
+      snPhi = sin(phiSect);
+      csPhi = cos(phiSect);
+      xcSec = xc * csPhi + yc * snPhi; // recalculate circle center in the sector frame
+    }
+    float cstalp = (param::RowX[iRow] - xcSec) / r;
+    if (fabsf(cstalp) > 1.f - sFloatEps) {
+      // track cannot reach this pad row
+      cstalp = std::copysign(1.f - sFloatEps, cstalp);
+    }
+    params.tglArr[iCl] = cstalp / sqrt((1 - cstalp) * (1 + cstalp)); // 1 / tan(acos(cstalp)) = cstalp / sqrt(1 - cstalp^2)
+
+    // In B+ the slope of q- should increase with x. Just look on q * B
+    if (params.qpt * mBz > 0) {
+      params.tglArr[iCl] *= -1.f;
+    }
+  }
+  // LOGF(info, "CompareToHelix: hMaxY(%f), hMinY(%f), hMaxZ(%f), hMinZ(%f). Max deviation allowed: y(%.2f), z(%.2f)", hMaxY, hMinY, hMaxZ, hMinZ, param::MaxDevHelixY, param::MaxDevHelixZ);
+  // LOGF(info, "New pt/Q (%f), old pt/Q (%f)", 1./params.qpt, 1./trk.qPt);
+  return fabsf(hMaxY - hMinY) < param::MaxDevHelixY && fabsf(hMaxZ - hMinZ) < param::MaxDevHelixZ;
+}
+
+bool TrackInterpolation::outlierFiltering(const TrackData& trk, TrackParams& params, const std::vector<TPCClusterResiduals>& clsRes) const
+{
+  if (clsRes.size() < mNMALong) {
+    LOG(debug) << "Skipping track with too few clusters for long moving average: " << clsRes.size();
+    return false;
+  }
+  float rmsLong = checkResiduals(trk, params, clsRes);
+  if (static_cast<float>(params.flagRej.count()) / clsRes.size() > mMaxRejFrac) {
+    LOG(debug) << "Skipping track with too many clusters rejected: " << static_cast<float>(params.flagRej.count()) / clsRes.size();
+    return false;
+  }
+  if (rmsLong > mMaxRMSLong) {
+    LOG(debug) << "Skipping track with too large RMS: " << rmsLong;
+    return false;
+  }
+  return true;
+}
+
+float TrackInterpolation::checkResiduals(const TrackData& trk, TrackParams& params, const std::vector<TPCClusterResiduals>& clsRes) const
+{
+  float rmsLong = 0.f;
+
+  int nCl = clsRes.size();
+  int iClFirst = 0;
+  int iClLast = nCl - 1;
+  int secStart = clsRes[0].sec;
+
+  // arrays with differences / abs(differences) of points to their neighbourhood, initialized to zero
+  std::array<float, param::NPadRows> yDiffLL{};
+  std::array<float, param::NPadRows> zDiffLL{};
+  std::array<float, param::NPadRows> absDevY{};
+  std::array<float, param::NPadRows> absDevZ{};
+
+  for (unsigned int iCl = 0; iCl < nCl; ++iCl) {
+    if (iCl < iClLast && clsRes[iCl].sec == secStart) {
+      continue;
+    }
+    // sector changed or last cluster reached
+    // now run estimators for all points in the same sector
+    int nClSec = iCl - iClFirst;
+    if (iCl == iClLast) {
+      ++nClSec;
+    }
+    diffToLocLine(nClSec, iClFirst, params.xTrk, params.dy, yDiffLL);
+    diffToLocLine(nClSec, iClFirst, params.xTrk, params.dz, zDiffLL);
+    iClFirst = iCl;
+    secStart = clsRes[iCl].sec;
+  }
+  // store abs deviations
+  int nAccY = 0;
+  int nAccZ = 0;
+  for (int iCl = nCl; iCl--;) {
+    if (fabsf(yDiffLL[iCl]) > param::sEps) {
+      absDevY[nAccY++] = fabsf(yDiffLL[iCl]);
+    }
+    if (fabsf(zDiffLL[iCl]) > param::sEps) {
+      absDevZ[nAccZ++] = fabsf(zDiffLL[iCl]);
+    }
+  }
+  if (nAccY < param::MinNumberOfAcceptedResiduals || nAccZ < param::MinNumberOfAcceptedResiduals) {
+    // mask all clusters
+    params.flagRej.set();
+    return 0.f;
+  }
+  // estimate rms on 90% of the smallest deviations
+  int nKeepY = static_cast<int>(.9 * nAccY);
+  int nKeepZ = static_cast<int>(.9 * nAccZ);
+  std::nth_element(absDevY.begin(), absDevY.begin() + nKeepY, absDevY.begin() + nAccY);
+  std::nth_element(absDevZ.begin(), absDevZ.begin() + nKeepZ, absDevZ.begin() + nAccZ);
+  float rmsYkeep = 0.f;
+  float rmsZkeep = 0.f;
+  for (int i = nKeepY; i--;) {
+    rmsYkeep += absDevY[i] * absDevY[i];
+  }
+  for (int i = nKeepZ; i--;) {
+    rmsZkeep += absDevZ[i] * absDevZ[i];
+  }
+  rmsYkeep = std::sqrt(rmsYkeep / nKeepY);
+  rmsZkeep = std::sqrt(rmsZkeep / nKeepZ);
+  if (rmsYkeep < param::sEps || rmsZkeep < param::sEps) {
+    LOG(warning) << "Too small RMS: " << rmsYkeep << "(y), " << rmsZkeep << "(z).";
+    params.flagRej.set();
+    return 0.f;
+  }
+  float rmsYkeepI = 1.f / rmsYkeep;
+  float rmsZkeepI = 1.f / rmsZkeep;
+  int nAcc = 0;
+  std::array<float, param::NPadRows> yAcc;
+  std::array<float, param::NPadRows> yDiffLong;
+  for (int iCl = 0; iCl < nCl; ++iCl) {
+    yDiffLL[iCl] *= rmsYkeepI;
+    zDiffLL[iCl] *= rmsZkeepI;
+    if (yDiffLL[iCl] * yDiffLL[iCl] + zDiffLL[iCl] * zDiffLL[iCl] > param::mMaxStdDevMA) {
+      params.flagRej.set(iCl);
+    } else {
+      yAcc[nAcc++] = params.dy[iCl];
+    }
+  }
+  if (nAcc > mNMALong) {
+    diffToMA(nAcc, yAcc, yDiffLong);
+    float average = 0.f;
+    float rms = 0.f;
+    for (int i = 0; i < nAcc; ++i) {
+      average += yDiffLong[i];
+      rms += yDiffLong[i] * yDiffLong[i];
+    }
+    average /= nAcc;
+    rmsLong = rms / nAcc - average * average;
+    rmsLong = (rmsLong > 0) ? std::sqrt(rmsLong) : 0.f;
+  }
+  return rmsLong;
+}
+
+void TrackInterpolation::diffToLocLine(const int np, int idxOffset, const std::array<float, param::NPadRows>& x, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffY) const
+{
+  // Calculate the difference between the points and the linear extrapolations from the neighbourhood.
+  // Nothing more than multiple 1-d fits at once. Instead of building 4 sums (x, x^2, y, xy), 4 * nPoints sums are calculated at once
+  // compare to TrackResiduals::fitPoly1() method
+
+  // adding one entry to the vectors saves an additional if statement when calculating the cumulants
+  std::vector<float> sumX1vec(np + 1);
+  std::vector<float> sumX2vec(np + 1);
+  std::vector<float> sumY1vec(np + 1);
+  std::vector<float> sumXYvec(np + 1);
+  auto sumX1 = &(sumX1vec[1]);
+  auto sumX2 = &(sumX2vec[1]);
+  auto sumY1 = &(sumY1vec[1]);
+  auto sumXY = &(sumXYvec[1]);
+
+  // accumulate sums for all points
+  for (int iCl = 0; iCl < np; ++iCl) {
+    int idx = iCl + idxOffset;
+    sumX1[iCl] = sumX1[iCl - 1] + x[idx];
+    sumX2[iCl] = sumX2[iCl - 1] + x[idx] * x[idx];
+    sumY1[iCl] = sumY1[iCl - 1] + y[idx];
+    sumXY[iCl] = sumXY[iCl - 1] + x[idx] * y[idx];
+  }
+
+  for (int iCl = 0; iCl < np; ++iCl) {
+    int iClLeft = iCl - mNMAShort;
+    int iClRight = iCl + mNMAShort;
+    if (iClLeft < 0) {
+      iClLeft = 0;
+    }
+    if (iClRight >= np) {
+      iClRight = np - 1;
+    }
+    int nPoints = iClRight - iClLeft;
+    if (nPoints < mNMAShort) {
+      continue;
+    }
+    float nPointsInv = 1.f / nPoints;
+    int iClLeftP = iClLeft - 1;
+    int iClCurrP = iCl - 1;
+    // extract sum from iClLeft to iClRight from cumulants, excluding iCl from the fit
+    float sX1 = sumX1[iClRight] - sumX1[iClLeftP] - (sumX1[iCl] - sumX1[iClCurrP]);
+    float sX2 = sumX2[iClRight] - sumX2[iClLeftP] - (sumX2[iCl] - sumX2[iClCurrP]);
+    float sY1 = sumY1[iClRight] - sumY1[iClLeftP] - (sumY1[iCl] - sumY1[iClCurrP]);
+    float sXY = sumXY[iClRight] - sumXY[iClLeftP] - (sumXY[iCl] - sumXY[iClCurrP]);
+    float det = sX2 - nPointsInv * sX1 * sX1;
+    if (fabsf(det) < 1e-12f) {
+      continue;
+    }
+    float slope = (sXY - nPointsInv * sX1 * sY1) / det;
+    float offset = nPointsInv * sY1 - nPointsInv * slope * sX1;
+    diffY[iCl + idxOffset] = y[iCl + idxOffset] - slope * x[iCl + idxOffset] - offset;
+  }
+}
+
+void TrackInterpolation::diffToMA(const int np, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffMA) const
+{
+  // Calculate
+  std::vector<float> sumVec(np + 1);
+  auto sum = &(sumVec[1]);
+  for (int i = 0; i < np; ++i) {
+    sum[i] = sum[i - 1] + y[i];
+  }
+  for (int i = 0; i < np; ++i) {
+    diffMA[i] = 0;
+    int iLeft = i - mNMALong;
+    int iRight = i + mNMALong;
+    if (iLeft < 0) {
+      iLeft = 0;
+    }
+    if (iRight >= np) {
+      iRight = np - 1;
+    }
+    int nPoints = iRight - iLeft;
+    if (nPoints < mNMALong) {
+      // this cannot happen, since at least mNMALong points are required as neighbours for this function to be called
+      continue;
+    }
+    float movingAverage = (sum[iRight] - sum[iLeft - 1] - (sum[i] - sum[i - 1])) / nPoints;
+    diffMA[i] = y[i] - movingAverage;
+  }
 }
 
 void TrackInterpolation::reset()
 {
   mTrackData.clear();
   mClRes.clear();
+  mTrackDataUnfiltered.clear();
+  mClResUnfiltered.clear();
   mGIDsSuccess.clear();
 }
