+++ Detectors/TPC/calibration/SpacePoints/src/TrackResiduals.cxx
@@ -200,12 +200,6 @@ void TrackResiduals::reset()
 }
 
 //______________________________________________________________________________
-void TrackResiduals::resetUnbinnedResiduals()
-{
-  mUnbinnedResiduals.clear();
-}
-
-//______________________________________________________________________________
 int TrackResiduals::getRowID(float x) const
 {
   int ix;
@@ -301,391 +295,6 @@ void TrackResiduals::setKernelType(KernelType kernel, float bwX, float bwP, floa
 /// processing functions
 ///
 ///////////////////////////////////////////////////////////////////////////////
-
-void TrackResiduals::prepareLocalResidualTrees()
-{
-  // prepare tree structure
-  mTmpFile = std::make_unique<TFile>(Form("%s.root", mLocalResFileName.c_str()), "recreate");
-  mTmpTree = std::make_unique<TTree>(mLocalResTreeName.c_str(), "TPC unbinned residuals");
-  mTmpTree->Branch(mLocalResBranchName.c_str(), &mUnbinnedResidualsPtr); // in the writer later on this should be one branch per sector?
-}
-
-void TrackResiduals::writeLocalResidualTreesToFile()
-{
-  // write trees with unbinned local residuals to file
-  mTmpFile->cd();
-  mTmpTree->Fill();
-  mTmpTree->Write();
-  mTmpTree.reset();
-  mTmpFile->Close();
-  mTmpFile.reset();
-}
-
-void TrackResiduals::loadInputFromFile()
-{
-  // open input file and access track data
-  mFileIn = std::make_unique<TFile>(mInputFileNameResiduals.data(), "open");
-  if (!mFileIn) {
-    LOG(error) << "input file could not be opened";
-    return;
-  }
-  mTreeInTracks = static_cast<TTree*>(mFileIn->Get("tracks"));
-  if (!mTreeInTracks) {
-    LOG(error) << "tree with track information not available in input file";
-    return;
-  }
-  mTreeInTracks->SetBranchAddress("tracks", &mTrackDataPtr);
-  mTreeInTracks->GetEntry(0);
-  // and access also cluster residuals
-  mTreeInClRes = static_cast<TTree*>(mFileIn->Get("residuals"));
-  if (!mTreeInClRes) {
-    LOG(error) << "tree with TPC cluster residuals not available in input file";
-    return;
-  }
-  mTreeInClRes->SetBranchAddress("residuals", &mClResPtr);
-  mTreeInClRes->GetEntry(0);
-}
-
-void TrackResiduals::setInputData(std::vector<TrackData>& trkData, std::vector<TPCClusterResiduals>& clResiduals)
-{
-  mTrackDataPtr = &trkData;
-  mClResPtr = &clResiduals;
-}
-
-void TrackResiduals::doOutlierRejection(bool writeToFile)
-{
-  // Receives the collected residuals from TrackInterpolation, does outlier filtering
-  // and writes the filtered residuals into a std::vector<UnbinnedResid> which will
-  // be send to the aggregator afterwards
-
-  if (!mIsInitialized) {
-    LOG(error) << "TrackResiduals class not initialized. Aborting";
-    return;
-  }
-
-  LOGF(info, "We got %lu tracks and %lu residuals", mTrackDataPtr->size(), mClResPtr->size());
-
-  const std::vector<TrackData>& trackData = *mTrackDataPtr;
-  const std::vector<TPCClusterResiduals>& clRes = *mClResPtr;
-
-  if (writeToFile) {
-    prepareLocalResidualTrees();
-  }
-
-  int nTracksValidated = 0;
-
-  // loop over tracks
-  for (const auto& trk : trackData) {
-    TrackParams params;
-    if (!validateTrack(trk, params) && mFilterOutliers) {
-      continue;
-    }
-    TrackData trkDataOut = trk; // this track will be forwarded to the aggregator, but we still need to adapt the cluster range reference
-    int nClValidated = 0;
-    trkDataOut.clIdx.setFirstEntry(mUnbinnedResiduals.size());
-    ++nTracksValidated;
-    int iRow = 0;
-    for (int iCl = 0; iCl < trk.clIdx.getEntries(); ++iCl) {
-      int clIdx = trk.clIdx.getFirstEntry() + iCl;
-      iRow += clRes[clIdx].dRow;
-      if (params.flagRej[iCl]) {
-        // skip masked cluster residual
-        continue;
-      }
-      ++nClValidated;
-      float tgPhi = clRes[clIdx].snp / std::sqrt((1.f - clRes[clIdx].snp) * (1.f + clRes[clIdx].snp));
-      mUnbinnedResiduals.emplace_back(clRes[clIdx].dy, clRes[clIdx].dz, tgPhi, clRes[clIdx].y, clRes[clIdx].z, iRow, clRes[clIdx].sec);
-    }
-    trkDataOut.clIdx.setEntries(nClValidated);
-    mTrackDataOut.push_back(trkDataOut);
-  }
-
-  LOGF(info, "Could validate %i tracks out of %lu", nTracksValidated, mTrackDataPtr->size());
-
-  if (writeToFile) {
-    // write to file for debugging
-    writeLocalResidualTreesToFile();
-  }
-}
-
-bool TrackResiduals::validateTrack(const TrackData& trk, TrackParams& params) const
-{
-  if (trk.clIdx.getEntries() < param::MinNCl) {
-    // no enough clusters for this track to be considered
-    LOG(debug) << "Skipping track with too few clusters: " << trk.clIdx.getEntries();
-    return false;
-  }
-
-  bool resHelix = compareToHelix(trk, params);
-  if (!resHelix) {
-    LOG(debug) << "Skipping track too far from helix approximation";
-    return false;
-  }
-  if (fabsf(params.qpt) > param::MaxQ2Pt) {
-    LOG(debug) << "Skipping track with too high q/pT: " << params.qpt;
-    return false;
-  }
-  if (!outlierFiltering(trk, params)) {
-    return false;
-  }
-  return true;
-}
-
-bool TrackResiduals::outlierFiltering(const TrackData& trk, TrackParams& params) const
-{
-  if (trk.clIdx.getEntries() < mNMALong) {
-    LOG(debug) << "Skipping track with too few clusters for long moving average: " << trk.clIdx.getEntries();
-    return false;
-  }
-  float rmsLong = checkResiduals(trk, params);
-  if (static_cast<float>(params.flagRej.count()) / trk.clIdx.getEntries() > mMaxRejFrac) {
-    LOG(debug) << "Skipping track with too many clusters rejected: " << static_cast<float>(params.flagRej.count()) / trk.clIdx.getEntries();
-    return false;
-  }
-  if (rmsLong > mMaxRMSLong) {
-    LOG(debug) << "Skipping track with too large RMS: " << rmsLong;
-    return false;
-  }
-  return true;
-}
-
-float TrackResiduals::checkResiduals(const TrackData& trk, TrackParams& params) const
-{
-  float rmsLong = 0.f;
-
-  int nCl = trk.clIdx.getEntries();
-  int idxOffset = trk.clIdx.getFirstEntry();
-  int iClFirst = 0;
-  int iClLast = nCl - 1;
-  int secStart = (*mClResPtr)[idxOffset].sec;
-
-  // arrays with differences / abs(differences) of points to their neighbourhood, initialized to zero
-  std::array<float, param::NPadRows> yDiffLL{};
-  std::array<float, param::NPadRows> zDiffLL{};
-  std::array<float, param::NPadRows> absDevY{};
-  std::array<float, param::NPadRows> absDevZ{};
-
-  for (int iCl = 0; iCl < nCl; ++iCl) {
-    if (iCl < iClLast && (*mClResPtr)[idxOffset + iCl].sec == secStart) {
-      continue;
-    }
-    // sector changed or last cluster reached
-    // now run estimators for all points in the same sector
-    int nClSec = iCl - iClFirst;
-    if (iCl == iClLast) {
-      ++nClSec;
-    }
-    diffToLocLine(nClSec, iClFirst, params.xTrk, params.dy, yDiffLL);
-    diffToLocLine(nClSec, iClFirst, params.xTrk, params.dz, zDiffLL);
-    iClFirst = iCl;
-    secStart = (*mClResPtr)[idxOffset + iCl].sec;
-  }
-  // store abs deviations
-  int nAccY = 0;
-  int nAccZ = 0;
-  for (int iCl = nCl; iCl--;) {
-    if (fabsf(yDiffLL[iCl]) > param::sEps) {
-      absDevY[nAccY++] = fabsf(yDiffLL[iCl]);
-    }
-    if (fabsf(zDiffLL[iCl]) > param::sEps) {
-      absDevZ[nAccZ++] = fabsf(zDiffLL[iCl]);
-    }
-  }
-  if (nAccY < param::MinNumberOfAcceptedResiduals || nAccZ < param::MinNumberOfAcceptedResiduals) {
-    // mask all clusters
-    params.flagRej.set();
-    return 0.f;
-  }
-  // estimate rms on 90% of the smallest deviations
-  int nKeepY = static_cast<int>(.9 * nAccY);
-  int nKeepZ = static_cast<int>(.9 * nAccZ);
-  std::nth_element(absDevY.begin(), absDevY.begin() + nKeepY, absDevY.begin() + nAccY);
-  std::nth_element(absDevZ.begin(), absDevZ.begin() + nKeepZ, absDevZ.begin() + nAccZ);
-  float rmsYkeep = 0.f;
-  float rmsZkeep = 0.f;
-  for (int i = nKeepY; i--;) {
-    rmsYkeep += absDevY[i] * absDevY[i];
-  }
-  for (int i = nKeepZ; i--;) {
-    rmsZkeep += absDevZ[i] * absDevZ[i];
-  }
-  rmsYkeep = std::sqrt(rmsYkeep / nKeepY);
-  rmsZkeep = std::sqrt(rmsZkeep / nKeepZ);
-  if (rmsYkeep < param::sEps || rmsZkeep < param::sEps) {
-    LOG(warning) << "Too small RMS: " << rmsYkeep << "(y), " << rmsZkeep << "(z).";
-    params.flagRej.set();
-    return 0.f;
-  }
-  float rmsYkeepI = 1.f / rmsYkeep;
-  float rmsZkeepI = 1.f / rmsZkeep;
-  int nAcc = 0;
-  std::array<float, param::NPadRows> yAcc;
-  std::array<float, param::NPadRows> yDiffLong;
-  for (int iCl = 0; iCl < nCl; ++iCl) {
-    yDiffLL[iCl] *= rmsYkeepI;
-    zDiffLL[iCl] *= rmsZkeepI;
-    if (yDiffLL[iCl] * yDiffLL[iCl] + zDiffLL[iCl] * zDiffLL[iCl] > param::mMaxStdDevMA) {
-      params.flagRej.set(iCl);
-    } else {
-      yAcc[nAcc++] = params.dy[iCl];
-    }
-  }
-  if (nAcc > mNMALong) {
-    diffToMA(nAcc, yAcc, yDiffLong);
-    float average = 0.f;
-    float rms = 0.f;
-    for (int i = 0; i < nAcc; ++i) {
-      average += yDiffLong[i];
-      rms += yDiffLong[i] * yDiffLong[i];
-    }
-    average /= nAcc;
-    rmsLong = rms / nAcc - average * average;
-    rmsLong = (rmsLong > 0) ? std::sqrt(rmsLong) : 0.f;
-  }
-  return rmsLong;
-}
-
-bool TrackResiduals::compareToHelix(const TrackData& trk, TrackParams& params) const
-{
-  std::array<float, param::NPadRows> residHelixY;
-  std::array<float, param::NPadRows> residHelixZ;
-
-  std::array<float, param::NPadRows> xLab;
-  std::array<float, param::NPadRows> yLab;
-  std::array<float, param::NPadRows> sPath;
-
-  float curvature = fabsf(trk.p[o2::track::ParLabels::kQ2Pt] * mBz * o2::constants::physics::LightSpeedCm2S * 1e-14f);
-  int clIdxFirst = trk.clIdx.getFirstEntry();
-  int nCl = trk.clIdx.getEntries();
-  int secFirst = (*mClResPtr)[clIdxFirst].sec;
-  float phiSect = (secFirst + .5f) * o2::constants::math::SectorSpanRad;
-  float snPhi = sin(phiSect);
-  float csPhi = cos(phiSect);
-  sPath[0] = 0.f;
-
-  int iRow = 0;
-  for (int iP = 0; iP < nCl; ++iP) {
-    iRow += (*mClResPtr)[clIdxFirst + iP].dRow;
-    float yTrk = (*mClResPtr)[clIdxFirst + iP].y;
-    // LOGF(info, "iRow(%i), yTrk(%f)", iRow, yTrk);
-    xLab[iP] = param::RowX[iRow];
-    if ((*mClResPtr)[clIdxFirst + iP].sec != secFirst) {
-      float phiSectCurrent = ((*mClResPtr)[clIdxFirst + iP].sec + .5f) * o2::constants::math::SectorSpanRad;
-      float cs = cos(phiSectCurrent - phiSect);
-      float sn = sin(phiSectCurrent - phiSect);
-      xLab[iP] = param::RowX[iRow] * cs - yTrk * sn;
-      yLab[iP] = yTrk * cs + param::RowX[iRow] * sn;
-    } else {
-      xLab[iP] = param::RowX[iRow];
-      yLab[iP] = yTrk;
-    }
-    // this is needed only later, but we retrieve it already now to save another loop
-    params.zTrk[iP] = (*mClResPtr)[clIdxFirst + iP].z;
-    params.xTrk[iP] = param::RowX[iRow];
-    params.dy[iP] = (*mClResPtr)[clIdxFirst + iP].dy;
-    params.dz[iP] = (*mClResPtr)[clIdxFirst + iP].dz;
-    // done retrieving values for later
-    if (iP > 0) {
-      float dx = xLab[iP] - xLab[iP - 1];
-      float dy = yLab[iP] - yLab[iP - 1];
-      float ds2 = dx * dx + dy * dy;
-      float ds = sqrt(ds2); // circular path (linear approximation)
-      // if the curvature of the track or the (approximated) chord length is too large the more exact formula is used:
-      // chord length = 2r * asin(ds/(2r))
-      // using the first two terms of the tailer expansion for asin(x) ~ x + x^3 / 6
-      if (ds * curvature > 0.05) {
-        ds *= (1.f + ds2 * curvature * curvature / 24.f);
-      }
-      sPath[iP] = sPath[iP - 1] + ds;
-    }
-  }
-  float xcSec = 0.f;
-  float ycSec = 0.f;
-  float r = 0.f;
-  fitCircle(nCl, xLab, yLab, xcSec, ycSec, r, residHelixY);
-  // LOGF(info, "Done with circle fit. nCl(%i), xcSec(%f), ycSec(%f), r(%f).", nCl, xcSec, ycSec, r);
-  /*
-  for (int i=0; i<nCl; ++i) {
-    LOGF(info, "i(%i), xLab(%f), yLab(%f).", i, xLab[i], yLab[i]);
-  }
-  */
-  // determine curvature
-  float phiI = TMath::ATan2(yLab[0], xLab[0]);
-  float phiF = TMath::ATan2(yLab[nCl - 1], xLab[nCl - 1]);
-  if (phiI < 0) {
-    phiI += o2::constants::math::TwoPI;
-  }
-  if (phiF < 0) {
-    phiF += o2::constants::math::TwoPI;
-  }
-  float dPhi = phiF - phiI;
-  float curvSign = -1.f;
-  if (dPhi > 0) {
-    if (dPhi < o2::constants::math::PI) {
-      curvSign = 1.f;
-    }
-  } else if (dPhi < -o2::constants::math::PI) {
-    curvSign = 1.f;
-  }
-  params.qpt = std::copysign(1.f / (r * mBz * o2::constants::physics::LightSpeedCm2S * 1e-14f), curvSign);
-
-  // calculate circle coordinates in the lab frame
-  float xc = xcSec * csPhi - ycSec * snPhi;
-  float yc = xcSec * snPhi + ycSec * csPhi;
-
-  std::array<float, 2> pol1Z;
-  fitPoly1(nCl, sPath, params.zTrk, pol1Z);
-
-  params.tgl = pol1Z[0];
-
-  // max deviations in both directions from helix fit in y and z
-  float hMinY = 1e9f;
-  float hMaxY = -1e9f;
-  float hMinZ = 1e9f;
-  float hMaxZ = -1e9f;
-  // extract residuals in Z and fill track slopes in sector frame
-  int secCurr = secFirst;
-  iRow = 0;
-  for (int iCl = 0; iCl < nCl; ++iCl) {
-    iRow += (*mClResPtr)[clIdxFirst + iCl].dRow;
-    float resZ = params.zTrk[iCl] - (pol1Z[1] + sPath[iCl] * pol1Z[0]);
-    residHelixZ[iCl] = resZ;
-    if (resZ < hMinZ) {
-      hMinZ = resZ;
-    }
-    if (resZ > hMaxZ) {
-      hMaxZ = resZ;
-    }
-    if (residHelixY[iCl] < hMinY) {
-      hMinY = residHelixY[iCl];
-    }
-    if (residHelixY[iCl] > hMaxY) {
-      hMaxY = residHelixY[iCl];
-    }
-    int sec = (*mClResPtr)[clIdxFirst + iCl].sec;
-    if (sec != secCurr) {
-      secCurr = sec;
-      phiSect = (.5f + sec) * o2::constants::math::SectorSpanRad;
-      snPhi = sin(phiSect);
-      csPhi = cos(phiSect);
-      xcSec = xc * csPhi + yc * snPhi; // recalculate circle center in the sector frame
-    }
-    float cstalp = (param::RowX[iRow] - xcSec) / r;
-    if (fabsf(cstalp) > 1.f - sFloatEps) {
-      // track cannot reach this pad row
-      cstalp = std::copysign(1.f - sFloatEps, cstalp);
-    }
-    params.tglArr[iCl] = cstalp / sqrt((1 - cstalp) * (1 + cstalp)); // 1 / tan(acos(cstalp)) = cstalp / sqrt(1 - cstalp^2)
-
-    // In B+ the slope of q- should increase with x. Just look on q * B
-    if (params.qpt * mBz > 0) {
-      params.tglArr[iCl] *= -1.f;
-    }
-  }
-  // LOGF(info, "CompareToHelix: hMaxY(%f), hMinY(%f), hMaxZ(%f), hMinZ(%f). Max deviation allowed: y(%.2f), z(%.2f)", hMaxY, hMinY, hMaxZ, hMinZ, param::MaxDevHelixY, param::MaxDevHelixZ);
-  // LOGF(info, "New pt/Q (%f), old pt/Q (%f)", 1./params.qpt, 1./trk.qPt);
-  return fabsf(hMaxY - hMinY) < param::MaxDevHelixY && fabsf(hMaxZ - hMinZ) < param::MaxDevHelixZ;
-}
 
 //______________________________________________________________________________
 void TrackResiduals::processSectorResiduals(int iSec)
@@ -1422,90 +1031,6 @@ double TrackResiduals::getKernelWeight(std::array<double, 3> u2vec) const
 ///
 ///////////////////////////////////////////////////////////////////////////////
 
-void TrackResiduals::diffToMA(const int np, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffMA) const
-{
-  // Calculate
-  std::vector<float> sumVec(np + 1);
-  auto sum = &(sumVec[1]);
-  for (int i = 0; i < np; ++i) {
-    sum[i] = sum[i - 1] + y[i];
-  }
-  for (int i = 0; i < np; ++i) {
-    diffMA[i] = 0;
-    int iLeft = i - mNMALong;
-    int iRight = i + mNMALong;
-    if (iLeft < 0) {
-      iLeft = 0;
-    }
-    if (iRight >= np) {
-      iRight = np - 1;
-    }
-    int nPoints = iRight - iLeft;
-    if (nPoints < mNMALong) {
-      // this cannot happen, since at least mNMALong points are required as neighbours for this function to be called
-      continue;
-    }
-    float movingAverage = (sum[iRight] - sum[iLeft - 1] - (sum[i] - sum[i - 1])) / nPoints;
-    diffMA[i] = y[i] - movingAverage;
-  }
-}
-
-void TrackResiduals::diffToLocLine(const int np, int idxOffset, const std::array<float, param::NPadRows>& x, const std::array<float, param::NPadRows>& y, std::array<float, param::NPadRows>& diffY) const
-{
-  // Calculate the difference between the points and the linear extrapolations from the neighbourhood.
-  // Nothing more than multiple 1-d fits at once. Instead of building 4 sums (x, x^2, y, xy), 4 * nPoints sums are calculated at once
-  // compare to TrackResiduals::fitPoly1() method
-
-  // adding one entry to the vectors saves an additional if statement when calculating the cumulants
-  std::vector<float> sumX1vec(np + 1);
-  std::vector<float> sumX2vec(np + 1);
-  std::vector<float> sumY1vec(np + 1);
-  std::vector<float> sumXYvec(np + 1);
-  auto sumX1 = &(sumX1vec[1]);
-  auto sumX2 = &(sumX2vec[1]);
-  auto sumY1 = &(sumY1vec[1]);
-  auto sumXY = &(sumXYvec[1]);
-
-  // accumulate sums for all points
-  for (int iCl = 0; iCl < np; ++iCl) {
-    int idx = iCl + idxOffset;
-    sumX1[iCl] = sumX1[iCl - 1] + x[idx];
-    sumX2[iCl] = sumX2[iCl - 1] + x[idx] * x[idx];
-    sumY1[iCl] = sumY1[iCl - 1] + y[idx];
-    sumXY[iCl] = sumXY[iCl - 1] + x[idx] * y[idx];
-  }
-
-  for (int iCl = 0; iCl < np; ++iCl) {
-    int iClLeft = iCl - mNMAShort;
-    int iClRight = iCl + mNMAShort;
-    if (iClLeft < 0) {
-      iClLeft = 0;
-    }
-    if (iClRight >= np) {
-      iClRight = np - 1;
-    }
-    int nPoints = iClRight - iClLeft;
-    if (nPoints < mNMAShort) {
-      continue;
-    }
-    float nPointsInv = 1.f / nPoints;
-    int iClLeftP = iClLeft - 1;
-    int iClCurrP = iCl - 1;
-    // extract sum from iClLeft to iClRight from cumulants, excluding iCl from the fit
-    float sX1 = sumX1[iClRight] - sumX1[iClLeftP] - (sumX1[iCl] - sumX1[iClCurrP]);
-    float sX2 = sumX2[iClRight] - sumX2[iClLeftP] - (sumX2[iCl] - sumX2[iClCurrP]);
-    float sY1 = sumY1[iClRight] - sumY1[iClLeftP] - (sumY1[iCl] - sumY1[iClCurrP]);
-    float sXY = sumXY[iClRight] - sumXY[iClLeftP] - (sumXY[iCl] - sumXY[iClCurrP]);
-    float det = sX2 - nPointsInv * sX1 * sX1;
-    if (fabsf(det) < 1e-12f) {
-      continue;
-    }
-    float slope = (sXY - nPointsInv * sX1 * sY1) / det;
-    float offset = nPointsInv * sY1 - nPointsInv * slope * sX1;
-    diffY[iCl + idxOffset] = y[iCl + idxOffset] - slope * x[iCl + idxOffset] - offset;
-  }
-}
-
 float TrackResiduals::fitPoly1Robust(std::vector<float>& x, std::vector<float>& y, std::array<float, 2>& res, std::array<float, 3>& err, float cutLTM) const
 {
   // robust pol1 fit, modifies input arrays order
@@ -1782,7 +1307,7 @@ float TrackResiduals::getMAD2Sigma(std::vector<float> data) const
   return k * medianOfAbsDeviations;
 }
 
-void TrackResiduals::fitCircle(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, float& xc, float& yc, float& r, std::array<float, param::NPadRows>& residHelixY) const
+void TrackResiduals::fitCircle(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, float& xc, float& yc, float& r, std::array<float, param::NPadRows>& residHelixY)
 {
   // this fast algebraic circle fit is described here:
   // https://dtcenter.org/met/users/docs/write_ups/circle_fit.pdf
@@ -1832,7 +1357,7 @@ void TrackResiduals::fitCircle(int nCl, std::array<float, param::NPadRows>& x, s
   //printf("r = %.4f m, xc = %.4f, yc = %.4f\n", r/100.f, xc, yc);
 }
 
-bool TrackResiduals::fitPoly1(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, std::array<float, 2>& res) const
+bool TrackResiduals::fitPoly1(int nCl, std::array<float, param::NPadRows>& x, std::array<float, param::NPadRows>& y, std::array<float, 2>& res)
 {
   // fit a straight line y = ax + b to a given set of points (x,y)
   // no measurement errors assumed, no fit errors calculated
